migration_tool/code_parsing/src/parser.py
from typing import Union, List, Optional
from aws_cdk import Stack
from constructs import Construct
from codegen import Codebase

class CDKStackParser:

    def initialize_parser(self, directory: str):
        """
        Initialize the CDK parser with the given directory

        Args:
            directory: The directory to initialize the parser with
        """
        # Implementation will go here later
        codebase = Codebase(directory)
        return codebase
        raise NotImplementedError()
